requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
databricks-sdk>=0.102
httpx>=0.27
pydantic>=2.0
